fix: calls -> rename top-bar folder to tool-bar (since it switched places)
