+++ clients/web/src/components/calls/tool-bar/calls-clear-button.tsx
