+++ clients/web/src/components/calls/tool-bar/calls-scale-slider.tsx
